+++ umami/preprocessing.py
@@ -90,6 +90,13 @@ def get_parser():
     )
 
     parser.add_argument(
+        "--hybrid_validation",
+        action="store_true",
+        help="""When running the resampling, by giving this option, the hybrid
+        validation sample is resampled and not the training sample.""",
+    )
+
+    parser.add_argument(
         "--flavour",
         nargs="+",
         default=None,
@@ -139,21 +146,26 @@ if __name__ == "__main__":
     # Check for resampling
     elif args.resampling:
 
+        # Check if hybrid validation sample should be produced and set the option in
+        # the config
+        config.sampling["use_validation_samples"] = args.hybrid_validation
+
         # Copy config to output directory
         config.copy_to_out_dir("resampling")
 
         # Check the method which should be used for resampling
         if config.sampling["method"] == "count":
-            sampler = upt.UnderSampling(config)
+            sampler = upt.UnderSampling(config=config)
 
         elif config.sampling["method"] == "pdf":
-            sampler = upt.PDFSampling(config, flavour=args.flavour)
+            sampler = upt.PDFSampling(config=config, flavour=args.flavour)
 
         elif config.sampling["method"] == "importance_no_replace":
-            sampler = upt.UnderSamplingNoReplace(config)
+            sampler = upt.UnderSamplingNoReplace(config=config)
 
         elif config.sampling["method"] == "weighting":
-            sampler = upt.Weighting(config)
+            sampler = upt.Weighting(config=config)
+
         else:
             raise ValueError(
                 f'{config.sampling["method"]} as sampling method is not supported!'
@@ -161,6 +173,9 @@ if __name__ == "__main__":
 
         # Run the sampling with the selected method
         sampler.Run()
+
+        # Set the option back to False to ensure correct naming
+        config.sampling["use_validation_samples"] = False
 
     # Calculate the scale dicts of the previous resampled files
     elif args.scaling: